requirements.txt
absl-py==1.0.0
appdirs==1.4.4
argcomplete==1.12.3
argh==0.26.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==1.4.0
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
bcrypt==3.2.0
Beaker==1.11.0
beautifulsoup4==4.10.0
bleach==4.1.0
boto==2.49.0
boto3==1.21.6
botocore==1.24.6
bottle==0.12.19
bypy==1.7.12
CacheControl==0.12.6
cachetools==5.0.0
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
cffi==1.15.0
chardet==4.0.0
click==8.0.4
cmd2==2.3.3
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==36.0.1
cycler==0.10.0
Cython==0.29.28
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
distlib==0.3.4
distro==1.7.0
docutils==0.16
entrypoints==0.4
executing==0.8.3
fail2ban==0.11.2
Flask==2.0.2
flatbuffers==2.0.6
fonttools==4.29.1
future==0.18.2
gast==0.3.3
google-auth==1.34.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.43.2
h5py==3.6.0
html5lib==1.1
idna==3.3
imageio==2.16.0
importlib-metadata==4.8.1
inflect==5.4.0
inflection==0.5.1
ipykernel==6.9.1
ipython==8.1.0
ipython_genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.1
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keras @ file:///build/python-keras/src/keras-2.8.0-py2.py3-none-any.whl
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
librosa==0.8.1
llvmlite==0.39.0.dev0+42.gdf27949
lockfile==0.12.2
logutils==0.3.5
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
more-itertools==8.10.0
msgpack==1.0.3
multiprocess==0.70.12.2
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
notebook==6.4.8
numba==0.56.0.dev0+321.g59fa2e335
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
ordered-set==4.0.2
packaging==20.9
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.2
pathtools==0.1.2
pecan==1.4.1
pep517==0.12.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.1
ply==3.11
prettytable==2.2.1
progress==1.6
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyalpm==0.10.6
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pybind11==2.9.1
pycparser==2.21
pycuda==2021.1
pydicom==2.2.2
pydot==1.4.2
Pygments==2.11.2
pyinotify==0.9.6
pyOpenSSL==21.0.0
pyparsing==3.0.1
pyperclip==1.8.2
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyrsistent==0.18.1
python-dateutil==2.8.2
pytoml==0.1.21
pytools==2022.1
pytz==2021.3
PyWavelets==1.2.0
pyxdg==0.27
PyYAML==6.0
pyzmq==22.3.0
rados==2.0.0
ranger-fm==1.9.3
rbd==2.0.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
resolvelib==0.5.5
retrying==1.3.3
rgw==2.0.0
rsa==4.8
s3transfer==0.5.1
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
Send2Trash==1.8.0
sequitur-g2p==1.0.1668.21
shadowsocks==2.6.12
six==1.16.0
sounddevice==0.4.2
soupsieve==2.3.1
stack-data==0.2.0
systemd-python==234
TBB==0.2
team==1.0
tensorboard==2.8.0
tensorboard-data-server @ file:///tmp/tensorboard_data_server-0.7.0a0-py3-none-manylinux2010_x86_64.whl
tensorboard-plugin-wit @ file:///build/python-tensorboard_plugin_wit/src/tensorboard_plugin_wit-1.8.1/wit-pip/release/dist/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorboardX==2.5
tensorflow-estimator @ file:///build/python-tensorflow-estimator/src/estimator_pip/tensorflow_estimator-2.7.0-py2.py3-none-any.whl
tensorflow-gpu @ file:///build/tensorflow/src/tmpcuda/tensorflow_gpu-2.8.0-cp310-cp310-linux_x86_64.whl
tensorflow-io-gcs-filesystem==0.24.0
tensorflow-serving-api @ file:///tmp/pip/tensorflow_serving_api-2.7.0-py2.py3-none-any.whl
tensorflow-serving-api-gpu @ file:///tmp/pip/tensorflow_serving_api_gpu-2.7.0-py2.py3-none-any.whl
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
tf-estimator-nightly==2.9.0.dev2022032908
threadpoolctl==2.2.0
tifffile==2022.2.9
toml==0.10.2
tomli==2.0.0
torch==1.11.0rc3
torchfile==0.1.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.1.0
Unidecode==1.3.2
urllib3==1.26.8
visdom==0.1.8.9
waitress==2.0.0
watchdog==0.10.7
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.7
websocket-client==1.2.3
WebTest==3.0.0
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.13.3
xmltodict==0.12.0
yq==2.14.0
zipp==3.7.0
